fix (dev_spring_boot) project structure refactored
