+++ src/main/java/com/project/item/repository/ItemRepository.java
@@ -1,0 +1,10 @@
+package com.project.item.repository;
+
+import com.project.item.model.Item;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ItemRepository extends JpaRepository<Item, Long> {
+
+}
