--- src/main/java/com/project/repository/ItemRepository.java
@@ -1,10 +1,0 @@
-package com.project.repository;
-
-import com.project.model.Item;
-import org.springframework.data.jpa.repository.JpaRepository;;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ItemRepository extends JpaRepository<Item, Long> {
-
-}
