--- src/main/java/com/project/service/CrudService.java
@@ -1,19 +1,0 @@
-package com.project.service;
-
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.github.fge.jsonpatch.JsonPatchException;
-import com.github.fge.jsonpatch.mergepatch.JsonMergePatch;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-
-public interface CrudService<T> {
-    Page<T> getAll(Pageable pageable);
-
-    T get(Long id);
-
-    T create(T t);
-
-    T update(Long id, JsonMergePatch patch) throws JsonPatchException, JsonProcessingException;
-
-    void delete(Long id);
-}
