+++ src/main/java/com/project/utils/exceptionhandler/exceptions/ApplicationException.java
@@ -1,4 +1,4 @@
-package com.project.exceptionhandler.exceptions;
+package com.project.utils.exceptionhandler.exceptions;
 
 public class ApplicationException extends RuntimeException {
