+++ src/main/java/com/project/utils/exceptionhandler/exceptions/NoSuchElemException.java
@@ -1,4 +1,4 @@
-package com.project.exceptionhandler.exceptions;
+package com.project.utils.exceptionhandler.exceptions;
 
 public class NoSuchElemException extends ApplicationException {
